chore: extract parquet writer
